refactor: add switches as extentions to Views
