+++ lib/screens/home.dart
@@ -1,15 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:memory_lamp/models/enums/views.dart';
 import 'package:memory_lamp/providers/view_provider.dart';
-import 'package:memory_lamp/views/books.dart';
-import 'package:memory_lamp/views/home.dart';
-import 'package:memory_lamp/views/logout.dart';
-import 'package:memory_lamp/views/notification.dart';
-import 'package:memory_lamp/views/profile.dart';
-import 'package:memory_lamp/views/settings.dart';
 import 'package:memory_lamp/widgets/compound_widgets/ml_appbar.dart';
 import 'package:memory_lamp/widgets/compound_widgets/ml_drawer.dart';
-import 'package:memory_lamp/widgets/ml_text.dart';
 import 'package:provider/provider.dart';
 
 class HomeScreen extends StatelessWidget {
@@ -24,30 +17,10 @@ class HomeScreen extends StatelessWidget {
         drawer: const MLDrawer(),
         body: Consumer<ViewProvider>(
           builder: (BuildContext context, viewProvider, Widget? child) {
-            return _viewSwitch(viewProvider.selectedView);
+            return viewProvider.selectedView.view;
           },
         ),
       ),
     );
-  }
-
-  Widget _viewSwitch(Views view) {
-    switch (view) {
-      case Views.home:
-        return HomeView();
-      case Views.profile:
-        return ProfileView();
-      case Views.books:
-        return BooksView();
-      case Views.notification:
-        return NotificationView();
-      case Views.settings:
-        return SettingsView();
-      case Views.logout:
-        return LogoutView();
-
-      default:
-        return Center(child: MLText("View not found"));
-    }
   }
 }
